requirements.txt
appdirs==1.4.3
cffi==1.10.0
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
pycparser==2.17
pyparsing==2.2.0
sdl2-cffi==1.0.6
six==1.10.0
